+++ CHANGELOG.md
@@ -76,6 +76,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added option to set a custom callback for esp32 network driver
 `disconnected` events
 - Added `network:sta_status/0` to get the current connection state of the sta interface.
+- Added ESP32 `-DATOMVM_ELIXIR_SUPPORT=on` configuration option
 
 ### Changed
 
@@ -101,6 +102,8 @@ instead `badarg`.
 - Using a custom callback for STA disconnected events in esp32 network driver will stop automatic re-connect,
 allowing applications to use scan results or other means to decide when and where to connect.
 - ESP32 cmake build options are now also exposed in `idf.py menuconfig`.
+- ESP32 Elixir support is determined automatically from the offset of `boot.avm` in the partition
+table.
 
 ### Fixed